added render-onto-cube

+++ src/graphics_environment.h
@@ -67,7 +67,7 @@ private:
 	Ref<RenderPass> final_render_pass;
 	/* hawk tuah */
 	VkQueryPool query_pool = VK_NULL_HANDLE;
-	size_t query_offset = 0;
+	uint32_t query_offset = 0;
 	VkDescriptorPool descriptor_pool = VK_NULL_HANDLE;
 	VkDescriptorSetLayout scene_descriptor_set_layout = VK_NULL_HANDLE;
 	VkDescriptorSetLayout object_descriptor_set_layout = VK_NULL_HANDLE;
